move structs to data folder
