+++ chain-impl-mockchain/src/testing/data/mod.rs
@@ -2,10 +2,12 @@ mod address;
 mod keys;
 mod leader;
 mod stake_pool;
+mod vote;
 mod wallet;
 
 pub use address::*;
 pub use keys::KeysDb;
 pub use leader::*;
 pub use stake_pool::*;
+pub use vote::{CommitteeMember, CommitteeMembersManager};
 pub use wallet::*;
