+++ chain-impl-mockchain/src/testing/e2e/vote_private.rs
@@ -1,6 +1,8 @@
+use crate::testing::build_tally_decrypt_share;
+use crate::testing::data::CommitteeMembersManager;
 use crate::testing::VoteTestGen;
 use crate::{
-    certificate::{TallyDecryptShares, VotePlan},
+    certificate::VotePlan,
     fee::LinearFee,
     header::BlockDate,
     testing::{
@@ -11,58 +13,12 @@ use crate::{
     value::Value,
     vote::{Choice, PayloadType},
 };
-use chain_vote::{
-    committee::MemberSecretKey, MemberCommunicationKey, MemberPublicKey, MemberState, CRS,
-};
 use rand_chacha::ChaCha20Rng;
-use rand_core::{CryptoRng, RngCore, SeedableRng};
+use rand_core::SeedableRng;
 
 const ALICE: &str = "Alice";
 const STAKE_POOL: &str = "stake_pool";
 const VOTE_PLAN: &str = "fund1";
-
-struct CommitteeMembersManager {
-    members: Vec<CommitteeMember>,
-}
-
-struct CommitteeMember {
-    state: MemberState,
-}
-
-impl CommitteeMembersManager {
-    pub fn new(rng: &mut (impl RngCore + CryptoRng), threshold: usize, members_no: usize) -> Self {
-        let mut public_keys = Vec::new();
-        for _ in 0..members_no {
-            let private_key = MemberCommunicationKey::new(rng);
-            let public_key = private_key.to_public();
-            public_keys.push(public_key);
-        }
-
-        let crs = CRS::random(rng);
-
-        let mut members = Vec::new();
-        for i in 0..members_no {
-            let state = MemberState::new(rng, threshold, &crs, &public_keys, i);
-            members.push(CommitteeMember { state })
-        }
-
-        Self { members }
-    }
-
-    pub fn members(&self) -> &[CommitteeMember] {
-        &self.members
-    }
-}
-
-impl CommitteeMember {
-    pub fn public_key(&self) -> MemberPublicKey {
-        self.state.public_key()
-    }
-
-    pub fn secret_key(&self) -> &MemberSecretKey {
-        self.state.secret_key()
-    }
-}
 
 #[test]
 pub fn private_vote_cast_action_transfer_to_rewards_all_shares() {
